Added -keeptrigger to the RR

+++ src/commands/russianroulette.js
@@ -207,6 +207,12 @@ module.exports = {
                         args = args.filter(arrayItem => arrayItem !== invividualArgument);
                     } catch (e) {}
                 }
+                if (invividualArgument.toLowerCase().startsWith("-keeptrigger")) {
+                    try {
+                        guild.waitingForInteraction.data.russianroulette[message.channel.id].keepTrigger = true;
+                        args = args.filter(arrayItem => arrayItem !== invividualArgument);
+                    } catch (e) {}
+                }
             });
         }
 
@@ -252,7 +258,7 @@ module.exports = {
             components: [joinButton, cancelButton],
             failIfNotExists: false
         }, false).then(async msg => {
-            message.delete().catch(e => {
+            if(!guild.waitingForInteraction.data.russianroulette[message.channel.id].keepTrigger)message.delete().catch(e => {
                 MainLog.log(`Could not delete message [${message.id}] in [${message.channel.id}][${message.channel.guild.id}] Error : ${e}`.red); //Logging in file & console
                 if (typeof guild != "undefined" && guild.configurationManager.configuration.behaviour.logDiscordErrors && guild.logToChannel.initialized) guild.channelLog(`[ERR] Could not delete message [${message.id}] in [<#${message.channel.id}>(${message.channel.id})] Error : \`${e}\``); //Loggin in log channel if logDiscordErrors is set & the log channel is initialized
             });
